+++ sdks/python/run_pylint.sh
@@ -46,7 +46,7 @@ EXCLUDED_GENERATED_FILES=(
 "apache_beam/io/gcp/internal/clients/storage/storage_v1_client.py"
 "apache_beam/io/gcp/internal/clients/storage/storage_v1_messages.py"
 "apache_beam/coders/proto2_coder_test_messages_pb2.py"
-apache_beam/runners/api/*pb2*.py
+apache_beam/portability/runners/api/*pb2*.py
 )
 
 FILES_TO_IGNORE=""
